Fixed installer crash that was preventing launching of windscribe application. ws/client/desktop/client-desktop#465

+++ tools/build_all.py
@@ -36,6 +36,10 @@ BUILD_QMAKE_EXE = ""
 BUILD_MACDEPLOY = ""
 BUILD_INSTALLER_FILES = ""
 BUILD_SYMBOL_FILES = ""
+global NO_POST_CLEAN # bool
+global BUILD_APP # bool
+global BUILD_COM # bool
+global BUILD_INSTALLER # bool
 
 
 def ExtractAppVersion():
@@ -309,10 +313,6 @@ def BuildAuthHelperWin32(configdata, targetlist):
   with utl.PushDir() as current_wd:
     msg.Print("Current working dir: " + current_wd)
 
-    # create target location
-    tempTargetDir = BUILD_INSTALLER_FILES
-    # utl.CreateDirectory(tempTargetDir)
-
     # ws_com, ws_com_server, ws_proxy_stub
     for target in targetlist:
       component = configdata[target]
@@ -331,7 +331,7 @@ def BuildAuthHelperWin32(configdata, targetlist):
 
       # move necessary outputs to InstallerFiles (for deployment)
       srcTargetName = os.path.normpath(os.path.join(GetProjectFolder(c_subdir), "..", "Release", component["target"]))
-      destTargetName = os.path.join(tempTargetDir, component["target"])
+      destTargetName = os.path.join(BUILD_INSTALLER_FILES, component["target"])
       msg.Verbose("Moving " + srcTargetName + " -> " + destTargetName)
       utl.CopyFile(srcTargetName, destTargetName)
 
@@ -402,7 +402,8 @@ def BuildInstallerWin32(configdata, qt_root, msvc_root, crt_root):
   buildenv.update(iutl.GetVisualStudioEnvironment())
   buildenv.update({ "CL" : "/MP" })
   BuildComponent(installer_info, False, qt_root, buildenv)
-  utl.RemoveFile(archive_filename)
+  if not NO_POST_CLEAN:
+    utl.RemoveFile(archive_filename)
   final_installer_name = os.path.normpath(os.path.join(os.getcwd(),
     "Windscribe_{}.exe".format(BUILD_APP_VERSION_STRINGS[0])))
   utl.RenameFile(os.path.normpath(os.path.join(BUILD_INSTALLER_FILES,
@@ -415,7 +416,8 @@ def BuildInstallerMac(configdata, qt_root):
   msg.Info("Zipping...")
   installer_info = configdata[configdata["installer"]["macos"]]
   archive_filename = os.path.normpath(os.path.join(ROOT_DIR, installer_info["subdir"], "installer", "resources", "windscribe.7z"))
-  utl.RemoveFile(archive_filename)
+  if not NO_POST_CLEAN:
+    utl.RemoveFile(archive_filename)
   iutl.RunCommand(["7z", "a", archive_filename,
                    os.path.join(BUILD_INSTALLER_FILES, "Windscribe.app"),
                    "-y", "-bso0", "-bsp2"])
@@ -469,30 +471,35 @@ def BuildAll():
       raise iutl.InstallError("CRT files not found.")
   # Prepare output.
   artifact_dir = os.path.join(ROOT_DIR, "build-exe")
-  utl.RemoveDirectory(artifact_dir)
+  if not NO_POST_CLEAN:
+    utl.RemoveDirectory(artifact_dir)
   temp_dir = iutl.PrepareTempDirectory("installer")
   global BUILD_INSTALLER_FILES, BUILD_SYMBOL_FILES
   if current_os == "macos":
     BUILD_INSTALLER_FILES = os.path.join(ROOT_DIR, "installer", "mac", "binaries")
   else:
     BUILD_INSTALLER_FILES = os.path.join(temp_dir, "InstallerFiles")
-  utl.CreateDirectory(BUILD_INSTALLER_FILES, True)
+  utl.CreateDirectory(BUILD_INSTALLER_FILES, False if NO_POST_CLEAN else True)
   if current_os == "win32": 
     BUILD_SYMBOL_FILES = os.path.join(temp_dir, "SymbolFiles")
-    utl.CreateDirectory(BUILD_SYMBOL_FILES, True)
+    utl.CreateDirectory(BUILD_SYMBOL_FILES, False if NO_POST_CLEAN else True)
   # Build the components.
   GenerateProtobuf()
   with utl.PushDir(temp_dir):
-    if configdata["targets"][current_os]:
-      BuildComponents(configdata, configdata["targets"][current_os], qt_root)
+    if BUILD_APP:
+      if configdata["targets"][current_os]:
+        BuildComponents(configdata, configdata["targets"][current_os], qt_root)
     if current_os == "win32":
-      BuildAuthHelperWin32(configdata, configdata["targets"]["authhelper"])
-      BuildInstallerWin32(configdata, qt_root, msvc_root, crt_root)
+      if BUILD_COM:
+        BuildAuthHelperWin32(configdata, configdata["targets"]["authhelper"])
+      if BUILD_INSTALLER:
+        BuildInstallerWin32(configdata, qt_root, msvc_root, crt_root)
     elif current_os == "macos":
-      BuildInstallerMac(configdata, qt_root)
+      if BUILD_INSTALLER:
+        BuildInstallerMac(configdata, qt_root)
   # Copy artifacts.
   msg.Print("Installing artifacts...")
-  utl.CreateDirectory(artifact_dir, True)
+  utl.CreateDirectory(artifact_dir, False if NO_POST_CLEAN else True)
   if current_os == "macos":
     artifact_path = BUILD_INSTALLER_FILES
     installer_info = configdata[configdata["installer"]["macos"]]
@@ -507,13 +514,20 @@ def BuildAll():
     utl.CopyFile(filename, os.path.join(artifact_dir, filetitle))
     msg.HeadPrint("Ready: \"{}\"".format(filetitle))
   # Cleanup.
-  msg.Print("Cleaning temporary directory...")
-  utl.RemoveDirectory(temp_dir)
+  if not NO_POST_CLEAN:
+    msg.Print("Cleaning temporary directory...")
+    utl.RemoveDirectory(temp_dir)
 
 
 if __name__ == "__main__":
   start_time = time.time()
   current_os = utl.GetCurrentOS()
+
+  NO_POST_CLEAN = "--no-clean" in sys.argv
+  BUILD_APP = not ("--no-app" in sys.argv)
+  BUILD_COM = not ("--no-com" in sys.argv)
+  BUILD_INSTALLER = not ("--no-installer" in sys.argv)
+
   if current_os not in BUILD_OS_LIST:
     msg.Print("{} is not needed on {}, skipping.".format(BUILD_TITLE, current_os))
     sys.exit(0)
